requirements.txt
Flask==2.0.2
werkzeug==2.0.2
PyMuPDF==1.21.0
python-docx==0.8.11
matplotlib==3.5.2
